# Conflicts:
#	Sistema-de-Gestao-Consultas/src/br/edu/ufersa/controlConsult/SistemaDeGestaoHospitalar.java

+++ Sistema-de-Gestao-Consultas/src/br/edu/ufersa/controlConsult/gui/FormPessoa.java
@@ -85,6 +85,7 @@ public class FormPessoa extends javax.swing.JFrame {
                     pessoa_jPanel.setVisible(false);
                     medico_jPanel.setVisible(false);
                     paciente_jPanel.setVisible(false);
+                    botoes_jPanel.setVisible(false);
                 }
                 break;
             default:
@@ -817,6 +818,7 @@ public class FormPessoa extends javax.swing.JFrame {
         }
         Especialidade especialidade = especialidesMap.get(espField_jComboBox.getSelectedItem());
         Medico m = new Medico(crm, cargaHoraria, especialidade);
+        m.setPessoa(pessoa);
         pessoa.setMedico(m);
     }
 
@@ -970,10 +972,12 @@ public class FormPessoa extends javax.swing.JFrame {
     }//GEN-LAST:event_chFieldActionPerformed
 
     private void search_jButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_search_jButtonActionPerformed
+
         try {
             limpaFormulario();
             pessoa = Pessoa.findByCPF(BuscaCpf_textField.getText());
             preencherFormularioPessoa();
+            botoes_jPanel.setVisible(true);
         } catch (NoResultException ex) {
             JOptionPane.showMessageDialog(null, "Ninguém encontrado.");
         } finally {
@@ -995,10 +999,13 @@ public class FormPessoa extends javax.swing.JFrame {
 
     private void delete_jToggleButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_delete_jToggleButtonActionPerformed
         if (pessoa != null) {
-            pessoa.delete();
-            limpaFormulario();
-            pessoa = null;
-            atualizarContextoJanela();
+            int reply = JOptionPane.showConfirmDialog(null, "Deseja Realmente apagar a pessoa " + pessoa.getNome() + "?", "Confirmação", JOptionPane.YES_NO_OPTION);
+            if(reply == JOptionPane.YES_OPTION){
+                pessoa.delete();
+                limpaFormulario();
+                pessoa = null;
+                atualizarContextoJanela();
+            }            
         }
     }//GEN-LAST:event_delete_jToggleButtonActionPerformed
